Updated the view and controller to enable searching
Created a search object, the intent is that as we add more search criteria (e.g date/time filters) they should be able to be submitted through the same form. This way all the info can be brought together in the search object. 

The search() method can then be expanded to perform more complex searches with multiple filters depending on user's input