Create the header and relevent styling
In the individual html.hbs files the page is specified in the 'navbar' partial, which the heading in base.html.hbs then uses (in the sass) to highlight the page the user is currently on.

This is slightly messy admitidly, but with the way the system is constructed I can't think of a cleaner way of doing this.