Added tests to check update, findAllExceptOne and the new form view.
Update test creates event, then uses update method on it and checks
whether the original event has the changes made to it.
Find all except one tests whether a created venue is not found in the
list given by the findAllExceptOne method.
Form test checks that on a get request, the correct view is shown to the
user.