Split up date and time, allowing an event not to have a time, added small validation to update form and to Event class.
Date and time are now separated, as events are not required to have a
time set. The home page now shows the time of the event, if an event has
one, and when updating there are separate forms for selecting the date
and time of the event. The form will also now fail and display an error
message if the name or date is left out, and revert to the default data
of the event.
This does mean another constructor has been added and so any tests where
an event is created needs another date parameter added. Fortunately you
can just use the same date object twice and it seems to be parsed
correctly for both date and time.