Implemented co-ordinate geocoding for venues with tests.
Split address and postcode into two variables to allow for easier
validation in future. Added the extra variable to current tests' venues.
Currently the co-ordinates are created on either constructing a venue or
using the setCoords() method (will be needed if venue address is
updated).

My reasoning for putting the request in the venue entity class is that I
believed it would be called the least from there since it would only be
called if completely necessary (i.e. when a venue is created or address
of venue is updated). I might be wrong and am open to moving it
elsewhere. I was also unsure where exactly to put the tests for it so I
created a new test class.