Fixed venues deletion. Issue #49
Venues without any events can now be deleted using the button on their
individual page. However, when trying to delete a venue that is linked
to an event, a constraint error should appear. (but it doesnt right now)
A message is registered in both cases. The message can be further used
for dialogues issue.